bug fix
git-svn-id: https://subversion.assembla.com/svn/smart-cs/trunk@8293 37136b1e-4cb1-4992-8394-573b48a9bdce

+++ source/extensions/cybertracker/org.wcs.smart.cybertracker/src/org/wcs/smart/cybertracker/export/CtJsonExportUtils.java
@@ -836,8 +836,11 @@ public class CtJsonExportUtils {
 					}
 					if (imgFile != null) {
 						Path target = targetDir.resolve(imgFile.getFileName());
-						Files.copy(imgFile, target);
-						filesToAdd.add(target);
+						if(!Files.exists(target)) {
+							//file may exist if the patrol cm uses same cm as incident model
+							Files.copy(imgFile, target);
+							filesToAdd.add(target);
+						}
 					}
 
 					AttributeOptionType op = new AttributeOptionType();
